requirements.txt
google-generativeai
python-dotenv
requests
pydantic
fastapi
uvicorn
feedparser
rich
pylatex